Add option to toggle sound
