+++ keymaps/keyboard-sounds.cson
@@ -1,0 +1,2 @@
+'atom-workspace':
+  'ctrl-alt-k': 'keyboard-sounds:toggle'
